+++ example/tutorial/step3/app.py
@@ -6,8 +6,7 @@ import os
 import secrets
 import signal
 
-from websockets.asyncio.connection import broadcast
-from websockets.asyncio.server import serve
+from websockets.asyncio.server import broadcast, serve
 
 from connect4 import PLAYER1, PLAYER2, Connect4